requirements.txt
rasterio
numpy
scipy
flask
flask-cors
Pillow
